added object based authentication logic

+++ delete_picture.php
@@ -1,6 +1,10 @@
 <?php
-    if ($_GET['picture']) {
-        unlink('./storage/images/'.$_GET['picture']);
-        header('location: index.php');
+    if (!$_SESSION['authenticated']) {
+        header('location: login.php');
+    } else {
+        if ($_GET['picture']) {
+            unlink('./storage/images/'.$_GET['picture']);
+            header('location: index.php');
+        }
     }
 ?>
